Face Recognition
This project will highlight the Face Recognition technique used to detect and
recognize faces of individuals where it is involved in capturing people faces from a
web camera. It involves candidates and trainers of CERTIFAI Online CDLE. The
individual’s images are categorized into two, which are the faces of trainers or
candidates. These images are then saved in the face database. The model thenshould be able to analyze face’s features and match it with the information in a database and identify who there are.